+++ src/core/error.rs
@@ -50,7 +50,7 @@ pub enum MonorailError {
     MissingArg(String),
     TaskCancelled,
     ChannelSend(String),
-    ChannelRecv(flume::RecvError),
+    ChannelRecv(String),
     Server(server::ServerError),
 }
 impl From<server::ServerError> for MonorailError {
@@ -58,13 +58,8 @@ impl From<server::ServerError> for MonorailError {
         MonorailError::Server(error)
     }
 }
-impl From<flume::RecvError> for MonorailError {
-    fn from(error: flume::RecvError) -> Self {
-        MonorailError::ChannelRecv(error)
-    }
-}
-impl<T> From<flume::SendError<T>> for MonorailError {
-    fn from(error: flume::SendError<T>) -> Self {
+impl<T> From<tokio::sync::mpsc::error::SendError<T>> for MonorailError {
+    fn from(error: tokio::sync::mpsc::error::SendError<T>) -> Self {
         MonorailError::ChannelSend(error.to_string())
     }
 }